Refactored methods in pd service
