+++ modules/command/dc/dc-convert-to-ot-json-command.js
@@ -11,6 +11,7 @@ class DcConvertToOtJsonCommand extends Command {
         this.commandExecutor = ctx.commandExecutor;
         this.config = ctx.config;
         this.web3 = ctx.web3;
+        this.importService = ctx.importService;
     }
 
     /**
@@ -24,7 +25,7 @@ class DcConvertToOtJsonCommand extends Command {
                 let document = JSON.parse(fs.readFileSync(documentPath, { encoding: 'utf-8' }));
 
                 if (!document.signature) {
-                    document = ImportUtilities.prepareDataset(document, this.config, this.web3);
+                    document = this.importService.prepareDataset(document);
                 }
 
                 fs.writeFileSync(documentPath, JSON.stringify(document));
