+++ test/modules/import-utilities.test.js
@@ -180,99 +180,6 @@ describe('Import utilities module ', () => {
         );
     });
 
-    it('Calculate the root hash on one permissioned data object', () => {
-        const originalObject = Utilities.copyObject(sample_data.permissioned_data_object);
-        const shuffledObject = Utilities.copyObject(sample_data.permissioned_data_object_shuffled);
-        const differentObject = Utilities.copyObject(sample_data.permissioned_data_object_2);
-
-        const originalRootHash = ImportUtilities.calculatePermissionedDataHash(originalObject);
-        const shuffledRootHash = ImportUtilities.calculatePermissionedDataHash(shuffledObject);
-        const differentRootHash = ImportUtilities.calculatePermissionedDataHash(differentObject);
-
-        assert(originalRootHash != null);
-        assert(shuffledRootHash != null);
-        assert(differentRootHash != null);
-
-        // Hashes should be 32 Bytes (64 characters) with 0x preceding the hash, so 66 characters
-        assert(typeof originalRootHash === 'string');
-        assert(typeof shuffledRootHash === 'string');
-        assert(typeof differentRootHash === 'string');
-
-        assert(originalRootHash.length === 66);
-        assert(shuffledRootHash.length === 66);
-        assert(differentRootHash.length === 66);
-
-        assert.equal(
-            originalRootHash,
-            shuffledRootHash,
-            'Permissioned data root hash for same object with attributes in different order!',
-        );
-
-        assert.notEqual(
-            originalRootHash,
-            differentRootHash,
-            'Permisssioned data root hash for different objects is the same!',
-        );
-    });
-
-    it('Calculating the root hash of an empty permissioned data object should throw an error', () => {
-        const testObject = {};
-
-        let errorHappened = false;
-        try {
-            ImportUtilities.calculatePermissionedDataHash(testObject);
-        } catch (e) {
-            errorHappened = true;
-            assert.equal(
-                e.message,
-                'Cannot calculate root hash of an empty object',
-                'Unexpected error received',
-            );
-        }
-
-        assert(errorHappened, 'calculatePermissionedDataHash did not throw an error!');
-    });
-
-    it('Calculate the public root hash of one graph', () => {
-        const originalGraph = Utilities.copyObject(sample_data.permissioned_data_graph['@graph']);
-        ImportUtilities.calculateGraphPermissionedDataHashes(originalGraph);
-
-        const shuffledGraph = Utilities.copyObject(sample_data.permissioned_data_graph_shuffled['@graph']);
-        ImportUtilities.calculateGraphPermissionedDataHashes(shuffledGraph);
-
-        const differentGraph = Utilities.copyObject(sample_data.permissioned_data_graph_2['@graph']);
-        ImportUtilities.calculateGraphPermissionedDataHashes(differentGraph);
-
-        const originalGraphRootHash = ImportUtilities.calculateGraphPublicHash(originalGraph);
-        const shuffledGraphRootHash = ImportUtilities.calculateGraphPublicHash(shuffledGraph);
-        const differentGraphRootHash = ImportUtilities.calculateGraphPublicHash(differentGraph);
-
-        assert(originalGraphRootHash != null);
-        assert(shuffledGraphRootHash != null);
-        assert(differentGraphRootHash != null);
-
-        // Hashes should be 32 Bytes (64 characters) with 0x preceding the hash, so 66 characters
-        assert(typeof originalGraphRootHash === 'string');
-        assert(typeof shuffledGraphRootHash === 'string');
-        assert(typeof differentGraphRootHash === 'string');
-
-        assert(originalGraphRootHash.length === 66);
-        assert(shuffledGraphRootHash.length === 66);
-        assert(differentGraphRootHash.length === 66);
-
-        assert.equal(
-            originalGraphRootHash,
-            shuffledGraphRootHash,
-            'Graph root hash for same object with attributes in different order!',
-        );
-
-        assert.notEqual(
-            originalGraphRootHash,
-            differentGraphRootHash,
-            'Graph root hash for different objects is the same!',
-        );
-    });
-
     it('Verify dataset signature', async () => {
         const testGraphCopy = Object.assign({}, sample_data.graph);
         const shuffledGraphCopy = Object.assign({}, sample_data.shuffledGraph);
@@ -291,36 +198,5 @@ describe('Import utilities module ', () => {
 
         assert.equal(signerOfOriginal, signerOfShuffled);
         assert.equal(signerOfShuffled, signingWallet.wallet);
-    });
-
-    it('Encoding verification', () => {
-        const permissionedObject = {
-            data: {
-                'urn:ot:object:product:batch:humidity': '19.7',
-                'urn:ot:object:product:batch:power_feeding': '85',
-                'urn:ot:object:product:batch:productId': 'urn:ot:object:actor:id:KakaxiSN687',
-                'urn:ot:object:product:batch:rainfall': '0.0',
-                'urn:ot:object:product:batch:solar_radiation': '0.0',
-                'urn:ot:object:product:batch:temperature': '22.0',
-                vocabularyType: 'urn:ot:object:batch',
-            },
-        };
-
-        const {
-            permissioned_data_original_length, permissioned_data_array_length, key,
-            encoded_data, permissioned_data_root_hash, encoded_data_root_hash,
-        } = ImportUtilities.encodePermissionedData(permissionedObject);
-
-        const result = ImportUtilities.validateAndDecodePermissionedData(
-            encoded_data,
-            key,
-            permissioned_data_array_length,
-            permissioned_data_original_length,
-        );
-
-        assert.equal(
-            Utilities.sortedStringify(permissionedObject.data),
-            Utilities.sortedStringify(result.permissionedData),
-        );
     });
 });
